feat(Home): Moved Home to views. Changed home content for menu
