+++ src/views/Home/Categorias.jsx
