+++ src/views/Home/Home.jsx
@@ -1,21 +1,21 @@
-import Expecial from "./Expecial"
-import Nav from "./Nav"
-
+import Categorias from "./Categorias"
+import Paginacion from "./Paginacion"
+import Card from "./Card"
 
 function Home() {
   return (
-    <div className="">
-       <Nav/>
-       <div className="flex flex-col items-center justify-center mt-16 ">
-        <div className="w-2/3 flex ">
-          <h1 className="text-[30px] font-bold text-red-600">
-            Especial del Dia
-          </h1>
+    <div>
+      <div className="flex justify-center">
+        <Categorias/>
+      </div>
+      <div className="flex justify-center">
+        <div className="flex  flex-wrap  rounded  mt-10 w-2/3 ">
+          <Card/>
         </div>
-         <div className="w-2/3">
-         <Expecial/>
-         </div>
-       </div>
+      </div>
+      <div className="flex justify-center m-5 items-center">
+        <Paginacion/>
+      </div>
     </div>
   )
 }
